Fixes #2628: Open bookmarks toggle emits an error
