+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 8.0.1 (Unreleased)
 
+- #2628: Open groupchats toggle emits an error
+- #2629: OMEMO button tooltip is wrong in private chat
 - #2634: Image previews not loading when not serving Converse locally
 
 ## 8.0.0 (2021-09-03)
